Shellrc modifications

+++ lib/cani/completions.rb
@@ -7,14 +7,14 @@ module Cani
       shw = Cani.api.browsers.reduce String.new do |acc, browser|
         versions = browser.versions.reverse.join(' ')
         acc +
-        "\ncomplete -f -c cani -n '__fish_cani_using_command show' -a '#{browser.abbr}' -d '#{browser.label}'" +
-        "\ncomplete -f -c cani -n '__fish_cani_showing_browser #{browser.abbr}' -a '#{versions}'"
+          "\ncomplete -f -c cani -n '__fish_cani_using_command show' -a '#{browser.abbr}' -d '#{browser.label}'" +
+          "\ncomplete -f -c cani -n '__fish_cani_showing_browser #{browser.abbr}' -a '#{versions}'"
       end
 
       use = Cani.api.features.reduce String.new do |acc, feature|
         description = feature.title.size > 40 ? feature.title[0..28] + '..' : feature.title
         acc +
-        "\ncomplete -f -c cani -n '__fish_cani_using_command use' -a '#{feature.name}' -d '#{description}'"
+          "\ncomplete -f -c cani -n '__fish_cani_using_command use' -a '#{feature.name}' -d '#{description}'"
       end
 
       tpl + shw + "\n" + use
@@ -26,8 +26,8 @@ module Cani
       indent    = 10
       versions  = Cani.api.browsers.reduce String.new do |acc, browser|
         acc + (' ' * (indent - 2)) + browser.abbr + ")\n" +
-        (' ' * indent) + "_arguments -C \"1: :(#{browser.versions.join(' ')})\"\n" +
-        (' ' * indent) + ";;\n"
+          (' ' * indent) + "_arguments -C \"1: :(#{browser.versions.join(' ')})\"\n" +
+          (' ' * indent) + ";;\n"
       end.strip
 
       tpl.gsub('{{names}}', Cani.api.browsers.map(&:abbr).join(' '))
@@ -41,8 +41,8 @@ module Cani
       indent    = 10
       versions  = Cani.api.browsers.reduce String.new do |acc, browser|
         acc + (' ' * (indent - 2)) + '"' + browser.abbr + "\")\n" +
-        (' ' * indent) + "COMPREPLY=($(compgen -W \"#{browser.versions.join(' ')}\" ${COMP_WORDS[COMP_CWORD]}))\n" +
-        (' ' * indent) + ";;\n"
+          (' ' * indent) + "COMPREPLY=($(compgen -W \"#{browser.versions.join(' ')}\" ${COMP_WORDS[COMP_CWORD]}))\n" +
+          (' ' * indent) + ";;\n"
       end.strip
 
       tpl.gsub('{{names}}', Cani.api.browsers.map(&:abbr).join(' '))
@@ -55,13 +55,8 @@ module Cani
       FileUtils.mkdir_p Cani.config.fish_comp_dir
       FileUtils.mkdir_p Cani.config.comp_dir
 
-      # write each completion file
-      File.open File.join(Cani.config.fish_comp_dir, 'cani.fish'), 'w' do |file|
-        file << generate_fish
-      end
-
-      %w[bash zsh].each do |shell|
-        File.open File.join(Cani.config.comp_dir, "_cani.#{shell}"), 'w' do |file|
+      %w[bash zsh fish].each do |shell|
+        File.open completion_path(shell), 'w' do |file|
           file << send("generate_#{shell}")
         end
       end
@@ -71,12 +66,8 @@ module Cani
     end
 
     def self.remove!
-      fish_comp = File.join Cani.config.fish_comp_dir, 'cani.fish'
-
-      File.unlink fish_comp if File.exist? fish_comp
-
-      %w[bash zsh].each do |shell|
-        shell_comp = File.join Cani.config.comp_dir, "_cani.#{shell}"
+      %w[bash zsh fish].each do |shell|
+        shell_comp = completion_path shell
 
         File.unlink shell_comp if File.exist? shell_comp
       end
@@ -85,12 +76,36 @@ module Cani
       delete_source_lines!
     end
 
+    def self.paths
+      puts "fish: #{completion_path(:fish)}"
+      puts "bash: #{completion_path(:bash)}"
+      puts "zsh:  #{completion_path(:zsh)}"
+    end
+
+    def self.completion_path(shell)
+      return File.join Cani.config.fish_comp_dir, 'cani.fish' if shell == :fish
+
+      File.join Cani.config.comp_dir, "_cani.#{shell}"
+    end
+
+    def self.compile_source_line(path)
+      return "test -e #{path}; and source #{path}" if path.end_with? '.fish'
+
+      "[ -f #{path} ] && source #{path}"
+    end
+
     def self.delete_source_lines!
+      return unless Cani.config.modify_shell_configs
+
       %w[bash zsh].each do |shell|
-        shellrc   = File.join Dir.home, ".#{shell}rc"
+        shellrc = File.join Dir.home, ".#{shell}rc"
+
+        next unless File.exist? shellrc
+
         lines     = File.read(shellrc).split "\n"
-        comp_path = File.join Cani.config.comp_dir, "_cani.#{shell}"
-        rm_idx    = lines.find_index { |l| l.match comp_path }
+        comp_path = completion_path shell
+        comp_src  = compile_source_line comp_path
+        rm_idx    = lines.find_index { |l| l.match comp_src }
 
         lines.delete_at rm_idx unless rm_idx.nil?
         File.write shellrc, lines.join("\n")
@@ -98,18 +113,19 @@ module Cani
     end
 
     def self.insert_source_lines!
-      %w[bash zsh].each do |shell|
-        shellrc   = File.join Dir.home, ".#{shell}rc"
-        lines     = File.read(shellrc).split "\n"
-        comp_path = File.join Cani.config.comp_dir, "_cani.#{shell}"
-        slidx     = lines.find_index { |l| l.match comp_path }
+      return unless Cani.config.modify_shell_configs
 
-        if slidx
-          lines[slidx] =
-            "#{lines[slidx][/^\s+/]}[ -f #{comp_path} ] && source #{comp_path}"
-        else
-          lines << "[ -f #{comp_path} ] && source #{comp_path}"
-        end
+      %w[bash zsh].each do |shell|
+        shellrc = File.join Dir.home, ".#{shell}rc"
+
+        next unless File.exist? shellrc
+
+        lines     = File.read(shellrc).split "\n"
+        comp_path = completion_path shell
+        comp_src  = compile_source_line comp_path
+
+        lines << comp_src \
+          unless lines.find_index { |l| l.match comp_src }
 
         File.write shellrc, lines.join("\n")
       end
